// lib/_TC/_SeD/enums/BrdBlm_TC_SeD_eFoamedPanelsOutline.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm_TC_SeD]
 * @author [APG] Angeli Paolo Giusto 
 * @version 0.1 APG 20231229
 * ----------------------------------------------------------------------------
 */

/**
 * Sagome delle sezioni dei pannellischiumati dei portoni sezionali
 */
export enum BrdBlm_TC_SeD_eFoamedPanelsOutline {

  MALE = "male",
  FEMALE = "female",
  STAVE = "stave",
  TOP_CUT = "top_cut",
  BOTTOM_CUT = "bottom_cut",

}